This is gross and a work in progress.
This takes the first steps towards moving the logic from showInputs() away from an HTML-based data orientation and towards a Javascript object-based data orientation.

Unfortunately I seem to have hit a spot with a bit of a predicament: I need to start using tuples of inputs so I can keep track of the quantity. The regrettable side effect of moving the inputs by distance in to tuples of { input, quantity } is that I lose access to inputs directly and have another level of abstraction.

Will deal with it later.